Move `Const.kt` to `core` package (#1789)
