+++ modules/core/arrow-core-data/src/test/kotlin/arrow/typeclasses/ConstTest.kt
@@ -1,10 +1,11 @@
 package arrow.typeclasses
 
+import arrow.core.Const
 import arrow.core.extensions.const.applicative.applicative
 import arrow.core.extensions.const.eq.eq
 import arrow.core.extensions.const.show.show
-import arrow.core.extensions.monoid
 import arrow.core.extensions.const.traverseFilter.traverseFilter
+import arrow.core.extensions.monoid
 import arrow.test.UnitSpec
 import arrow.test.laws.ApplicativeLaws
 import arrow.test.laws.EqLaws
