+++ modules/free/arrow-free-data/src/main/kotlin/arrow/free/FreeApplicative.kt
@@ -1,15 +1,15 @@
 package arrow.free
 
 import arrow.Kind
+import arrow.core.Const
+import arrow.core.ConstPartialOf
 import arrow.core.FunctionK
+import arrow.core.extensions.const.applicative.applicative
+import arrow.core.fix
+import arrow.core.value
 import arrow.higherkind
 import arrow.typeclasses.Applicative
-import arrow.typeclasses.Const
 import arrow.typeclasses.Monoid
-import arrow.typeclasses.fix
-import arrow.core.extensions.const.applicative.applicative
-import arrow.typeclasses.ConstPartialOf
-import arrow.typeclasses.value
 
 fun <F, G, A> FreeApplicativeOf<F, A>.foldMapK(f: FunctionK<F, G>, GA: Applicative<G>): Kind<G, A> =
   (this as FreeApplicative<F, A>).foldMap(f, GA)
