+++ modules/mtl/arrow-mtl-data/src/test/kotlin/arrow/mtl/KleisliTest.kt
@@ -1,6 +1,8 @@
 package arrow.mtl
 
 import arrow.Kind
+import arrow.core.Const
+import arrow.core.ConstPartialOf
 import arrow.core.Either
 import arrow.core.ForId
 import arrow.core.ForOption
@@ -8,17 +10,19 @@ import arrow.core.ForTry
 import arrow.core.Id
 import arrow.core.Option
 import arrow.core.Try
-import arrow.fx.ForIO
-import arrow.fx.IO
-import arrow.fx.extensions.io.applicativeError.attempt
-import arrow.fx.extensions.io.bracket.bracket
-import arrow.fx.mtl.kleisli.bracket.bracket
+import arrow.core.const
 import arrow.core.extensions.`try`.monadError.monadError
 import arrow.core.extensions.const.divisible.divisible
 import arrow.core.extensions.id.monad.monad
 import arrow.core.extensions.monoid
 import arrow.core.extensions.option.alternative.alternative
 import arrow.core.some
+import arrow.core.value
+import arrow.fx.ForIO
+import arrow.fx.IO
+import arrow.fx.extensions.io.applicativeError.attempt
+import arrow.fx.extensions.io.bracket.bracket
+import arrow.fx.mtl.kleisli.bracket.bracket
 import arrow.mtl.extensions.kleisli.alternative.alternative
 import arrow.mtl.extensions.kleisli.contravariant.contravariant
 import arrow.mtl.extensions.kleisli.divisible.divisible
@@ -30,13 +34,9 @@ import arrow.test.laws.ContravariantLaws
 import arrow.test.laws.DivisibleLaws
 import arrow.test.laws.MonadErrorLaws
 import arrow.typeclasses.Conested
-import arrow.typeclasses.Const
-import arrow.typeclasses.ConstPartialOf
 import arrow.typeclasses.Eq
 import arrow.typeclasses.conest
-import arrow.typeclasses.const
 import arrow.typeclasses.counnest
-import arrow.typeclasses.value
 import io.kotlintest.shouldBe
 
 class KleisliTest : UnitSpec() {
