+++ modules/ui/arrow-ui-data/src/test/kotlin/arrow/ui/SumTest.kt
@@ -1,6 +1,7 @@
 package arrow.ui
 
 import arrow.Kind
+import arrow.core.Const
 import arrow.core.ForId
 import arrow.core.Id
 import arrow.core.extensions.const.divisible.divisible
@@ -12,18 +13,17 @@ import arrow.core.extensions.id.functor.functor
 import arrow.core.extensions.id.hash.hash
 import arrow.core.extensions.monoid
 import arrow.core.fix
-import arrow.ui.extensions.sum.comonad.comonad
-import arrow.ui.extensions.sum.divisible.divisible
-import arrow.ui.extensions.sum.eq.eq
-import arrow.ui.extensions.sum.hash.hash
+import arrow.core.value
 import arrow.test.UnitSpec
 import arrow.test.laws.ComonadLaws
 import arrow.test.laws.DivisibleLaws
 import arrow.test.laws.HashLaws
-import arrow.typeclasses.Const
 import arrow.typeclasses.Eq
 import arrow.typeclasses.Hash
-import arrow.typeclasses.value
+import arrow.ui.extensions.sum.comonad.comonad
+import arrow.ui.extensions.sum.divisible.divisible
+import arrow.ui.extensions.sum.eq.eq
+import arrow.ui.extensions.sum.hash.hash
 import io.kotlintest.shouldBe
 
 class SumTest : UnitSpec() {
